+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 ### 💡 Enhancements 💡
 
 - Add `skip-get-modules` builder flag to support isolated environment executions (#6009)
+  - Skip unnecessary Go binary path validation when the builder is used with `skip-compilation` and `skip-get-modules` flags (#6026)
 
 ## v0.59.0 Beta